CineTome_back/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
passlib==1.7.4
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv
pydantic[email]
pydantic-settings>=2.0.0
gigachat==0.1.12
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
PyJWT==2.10.1
aiohttp==3.9.1
beautifulsoup4==4.12.0
fake-useragent==1.3.0